extrapolate all components
