+++ components/blocks/RecipeCodeBlockWithHighlight.tsx
@@ -1,0 +1,77 @@
+import Prism from 'prismjs';
+import React, { useEffect, useRef, useState } from 'react';
+
+import 'prismjs/plugins/line-numbers/prism-line-numbers';
+import 'prismjs/plugins/line-numbers/prism-line-numbers.css';
+import 'prism-themes/themes/prism-night-owl.css';
+import 'prismjs/plugins/line-highlight/prism-line-highlight';
+import 'prismjs/plugins/line-highlight/prism-line-highlight.css';
+
+import { CodeToolbar } from './RecipeCodeToolBar';
+
+interface CodeBlockProps {
+  value: any;
+  lang?: string;
+  children?: React.ReactNode;
+  highlightLines: string;
+}
+
+const CodeBlockWithHighlightLines = ({
+  value,
+  lang,
+  children,
+  highlightLines,
+}: CodeBlockProps) => {
+  const [tooltipVisible, setTooltipVisible] = useState(false);
+  const preRef = useRef<HTMLPreElement>(null);
+
+  useEffect(() => {
+    if (preRef.current) {
+      preRef.current.setAttribute('data-line', highlightLines);
+      // Apply Prism syntax highlighting to all elements under preRef
+      Prism.highlightAllUnder(preRef.current);
+    }
+  }, [highlightLines]);
+
+  const copyToClipboard = () => {
+    const codeToCopy = typeof children === 'string' ? children : value;
+    navigator.clipboard.writeText(codeToCopy).then(
+      () => {
+        setTooltipVisible(true);
+        setTimeout(() => setTooltipVisible(false), 1500);
+      },
+      (err) => {
+        console.error('Failed to copy code:', err);
+      }
+    );
+  };
+
+  return (
+    <div className="codeblock-container">
+      <div className="sticky top-0 z-30">
+        <CodeToolbar
+          lang={lang}
+          onCopy={copyToClipboard}
+          tooltipVisible={tooltipVisible}
+        />
+      </div>
+      <pre
+        ref={preRef}
+        className="line-numbers"
+        data-line={highlightLines} // Setting the lines to highlight
+        style={{
+          overflowX: 'hidden',
+          maxWidth: '100%',
+          whiteSpace: 'pre-wrap',
+          wordBreak: 'break-word',
+        }}
+      >
+        <code className={`language-${lang || 'jsx'}`}>
+          {typeof children === 'string' || children ? children : value}
+        </code>
+      </pre>
+    </div>
+  );
+};
+
+export default CodeBlockWithHighlightLines;
